remove deprecated alias for int

+++ tests/utils.py
@@ -9,7 +9,7 @@ def bin_mask_from_bb(mask_shape: Tuple[int, int], bbox: BBox):
     '''
       Draws a mask from a bounding box
     '''
-    return draw_rectangle(np.zeros(mask_shape, np.int),
+    return draw_rectangle(np.zeros(mask_shape, int),
                           bbox.upper_left_point,
                           bbox.bottom_right_point,
                           1)
